+++ scraper/src/index.ts
@@ -1,18 +1,19 @@
 import * as Queue from 'bull'
 import * as glob from 'glob'
 import { resolve } from 'path'
-import { Result } from './types/result'
+import { Result } from './sources/result'
 
 const scrapeAttributes = new Queue('scrape_attributes', process.env.REDIS_URL)
 const store = new Queue('store_results', process.env.REDIS_URL)
-const scrapeDetails = new Queue('scrape_detailed_attributes', process.env.REDIS_URL)
-const sources = glob.sync(resolve(__dirname, './source/**/*.js'))
-const sourceList: any = {}
+const sources = glob.sync(resolve(__dirname, './sources/sites/**/*.js'))
+let sourceList: any = {}
 
 sources.forEach(source => {
     const sourceModule = new (require(source).default)()
     sourceList[sourceModule.sourceName] = sourceModule
 })
+
+console.log('sourceList', sourceList);
 
 scrapeAttributes.process(1, require('./processors/scrape-attributes').bind({ sourceList }))
 scrapeAttributes
@@ -25,17 +26,11 @@ scrapeAttributes
 
         results.forEach((result: Result) => {
             store.add(result)
-            scrapeDetails.add({ source, link: result.link })
         })
     })
 
 store.on('active', (job: Queue.Job) => console.log('stored', job.data.link))
 store.process(500, require('./processors/store'))
-
-scrapeDetails.process(2, require('./processors/scrape-details').bind({store, sourceList}))
-scrapeDetails
-    .on('active', job => console.log('scrape details', job.data.source))
-    .on('failed', (job, err) => console.log('Error scraping details', err))
 
 function scrape() {
     console.log('--> start scraping')
@@ -43,7 +38,7 @@ function scrape() {
     for (let sourceName in sourceList) {
         const sourceModule = sourceList[sourceName]
         scrapeAttributes.add({ source: sourceModule.sourceName })
-        console.log('--> scrape', sourceModule.sourceName)
+        console.log('--> scrape', sourceName)
     }
 }