+++ scraper/src/sources/sites/thierry.ts
@@ -1,4 +1,4 @@
-import { HTMLSource } from '../types/html-source'
+import { HTMLSource } from '../html-source'
 
 export default class Thierry extends HTMLSource {
     public url = 'https://www.thierry-immobilier.fr/vente/appartement--maison'
@@ -16,20 +16,6 @@ export default class Thierry extends HTMLSource {
             selector: '> a',
             format($: CheerioStatic, link: CheerioStatic) {
                 return `https://www.thierry-immobilier.fr${$(link).attr('href')}`
-            }
-        }
-    ]
-
-    public richAttributes = [
-        {
-            type: 'photos',
-            selector: '.slick-track img',
-            format($: CheerioStatic, photos: CheerioStatic) {
-                const photoUrls = $(photos).map((index: number, photo: CheerioElement) => {
-                    return $(photo).attr('src').replace('styles/medium/public/', '')
-                }).get()
-
-                return photoUrls.filter((photo: string, index: number) => photoUrls.indexOf(photo) !== index)
             }
         }
     ]